requirements.txt
PyQt6
mss
Pillow
pywin32
